requirements.txt
telethon
python-dotenv
pymongo[srv]
